Separated Avatar logic into Avatar::Builder
